+++ CHANGELOG.md
@@ -18,7 +18,7 @@
 
 ### Removed
 
-## 0.6.0
+## 0.6.0-SNAPSHOT-1
 
 ### Added